initial refactor draft
