+++ doc/Makefile
@@ -3,6 +3,6 @@ all: doc
 %.1 %.1.html: %.md
 	ronn $^
 
-doc: kafkacat.1 kafkacat.1.html
+doc: kc.1 kc.1.html
 
 .PHONY: doc
