--- doc/kafkacat.1.html
@@ -1,179 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta http-equiv='content-type' value='text/html;charset=utf8'>
-  <meta name='generator' value='Ronn/v0.7.3 (http://github.com/rtomayko/ronn/tree/0.7.3)'>
-  <title>kafkacat(1) - generic producer and consumer for Apache Kafka</title>
-  <style type='text/css' media='all'>
-  /* style: man */
-  body#manpage {margin:0}
-  .mp {max-width:100ex;padding:0 9ex 1ex 4ex}
-  .mp p,.mp pre,.mp ul,.mp ol,.mp dl {margin:0 0 20px 0}
-  .mp h2 {margin:10px 0 0 0}
-  .mp > p,.mp > pre,.mp > ul,.mp > ol,.mp > dl {margin-left:8ex}
-  .mp h3 {margin:0 0 0 4ex}
-  .mp dt {margin:0;clear:left}
-  .mp dt.flush {float:left;width:8ex}
-  .mp dd {margin:0 0 0 9ex}
-  .mp h1,.mp h2,.mp h3,.mp h4 {clear:left}
-  .mp pre {margin-bottom:20px}
-  .mp pre+h2,.mp pre+h3 {margin-top:22px}
-  .mp h2+pre,.mp h3+pre {margin-top:5px}
-  .mp img {display:block;margin:auto}
-  .mp h1.man-title {display:none}
-  .mp,.mp code,.mp pre,.mp tt,.mp kbd,.mp samp,.mp h3,.mp h4 {font-family:monospace;font-size:14px;line-height:1.42857142857143}
-  .mp h2 {font-size:16px;line-height:1.25}
-  .mp h1 {font-size:20px;line-height:2}
-  .mp {text-align:justify;background:#fff}
-  .mp,.mp code,.mp pre,.mp pre code,.mp tt,.mp kbd,.mp samp {color:#131211}
-  .mp h1,.mp h2,.mp h3,.mp h4 {color:#030201}
-  .mp u {text-decoration:underline}
-  .mp code,.mp strong,.mp b {font-weight:bold;color:#131211}
-  .mp em,.mp var {font-style:italic;color:#232221;text-decoration:none}
-  .mp a,.mp a:link,.mp a:hover,.mp a code,.mp a pre,.mp a tt,.mp a kbd,.mp a samp {color:#0000ff}
-  .mp b.man-ref {font-weight:normal;color:#434241}
-  .mp pre {padding:0 4ex}
-  .mp pre code {font-weight:normal;color:#434241}
-  .mp h2+pre,h3+pre {padding-left:0}
-  ol.man-decor,ol.man-decor li {margin:3px 0 10px 0;padding:0;float:left;width:33%;list-style-type:none;text-transform:uppercase;color:#999;letter-spacing:1px}
-  ol.man-decor {width:100%}
-  ol.man-decor li.tl {text-align:left}
-  ol.man-decor li.tc {text-align:center;letter-spacing:4px}
-  ol.man-decor li.tr {text-align:right;float:right}
-  </style>
-</head>
-<!--
-  The following styles are deprecated and will be removed at some point:
-  div#man, div#man ol.man, div#man ol.head, div#man ol.man.
-
-  The .man-page, .man-decor, .man-head, .man-foot, .man-title, and
-  .man-navigation should be used instead.
--->
-<body id='manpage'>
-  <div class='mp' id='man'>
-
-  <div class='man-navigation' style='display:none'>
-    <a href="#NAME">NAME</a>
-    <a href="#SYNOPSIS">SYNOPSIS</a>
-    <a href="#DESCRIPTION">DESCRIPTION</a>
-    <a href="#OPTIONS">OPTIONS</a>
-  </div>
-
-  <ol class='man-decor man-head man head'>
-    <li class='tl'>kafkacat(1)</li>
-    <li class='tc'></li>
-    <li class='tr'>kafkacat(1)</li>
-  </ol>
-
-  <h2 id="NAME">NAME</h2>
-<p class="man-name">
-  <code>kafkacat</code> - <span class="man-whatis">generic producer and consumer for Apache Kafka</span>
-</p>
-
-<h2 id="SYNOPSIS">SYNOPSIS</h2>
-
-<p><code>kafkacat</code> -C [generic options] [-o offset] [-e] [-O] [-u]
-           -b brokers -t topic</p>
-
-<p><code>kafkacat</code> -P [generic options] [-z snappy | gzip] [-p -1] [file [...]]
-           -b brokers -t topic</p>
-
-<p><code>kafkacat</code> -L [generic options] [-t topic]</p>
-
-<h2 id="DESCRIPTION">DESCRIPTION</h2>
-
-<p><code>kafkacat</code> is a generic non-JVM producer and consumer for Apache Kafka 0.8,
-think of it as a netcat for Kafka.</p>
-
-<p>In producer mode ( -P ), <code>kafkacat</code> reads messages from stdin, delimited with a
-configurable delimeter and produces them to the provided Kafka cluster, topic
-and partition. In consumer mode ( -C ), <code>kafkacat</code> reads messages from a topic
-and partition and prints them to stdout using the configured message delimiter.</p>
-
-<p>If neither -P or -C are specified <code>kafkacat</code> attempts to figure out the mode
-automatically based on stdin/stdout tty types.</p>
-
-<p><code>kafkacat</code> also features a metadata list mode ( -L ), to display the
-current state of the Kafka cluster and its topics and partitions.</p>
-
-<h2 id="OPTIONS">OPTIONS</h2>
-
-<h3 id="Generic-Options">Generic Options</h3>
-
-<ul>
-<li><p><code>-t &lt;topic></code>
-Topic to consume from, produce to, or list.</p></li>
-<li><p><code>-b &lt;brokers></code>
-Comma separated list of broker(s) to bootstrap of the form
-"host[:port][,...]", e.g. "host1:9292,host2:9293".</p></li>
-<li><p><code>-p &lt;partition></code>
-Partition.</p></li>
-<li><p><code>-D &lt;delim></code>
-Message delimiter character: a-z.. | \r | \n | \t | \xNN [default: \n].</p></li>
-<li><p><code>-K &lt;delim></code>
-Key delimiter character: a-z.. | \r | \n | \t | \xNN.</p></li>
-<li><p><code>-c &lt;count></code>
-Limit message count. For the producer, exit after producing <code>&lt;count></code>
-messages. For the consumer, exit after consuming <code>&lt;count></code> messages.</p></li>
-<li><p><code>-X list|dump|&lt;prop>=&lt;val></code></p>
-
-<ul>
-<li><p><code>list</code> will list available librdkafka configuration properties.</p></li>
-<li><p><code>dump</code> will dump the configuration and exit.</p></li>
-<li><p><code>&lt;prop>=&lt;val></code> will set the property <var>prop</var> to value <var>val</var>. Properties
-prefixed with "topic" are applied as topic properties.</p></li>
-</ul>
-</li>
-<li><p><code>-d &lt;dbg></code>
-Enable librdkafka debugging. List of comma separated labels. See librdkafka
-documentation for supported values.</p></li>
-<li><p><code>-q</code>
-Quiet mode (set verbosity to 0).</p></li>
-<li><p><code>-v</code>
-Increase verbosity, can be called multiple time.</p></li>
-</ul>
-
-
-<h3 id="Producer-Options">Producer Options</h3>
-
-<ul>
-<li><p><code>-z snappy|gzip</code>
-Message compression [default: none].</p></li>
-<li><p><code>-p -1</code>
-Use random partitionner.</p></li>
-<li><p><code>-T</code>
-Output sent messages to stdout, acting like tee.</p></li>
-</ul>
-
-
-<h3 id="Consumer-Options">Consumer Options</h3>
-
-<ul>
-<li><p><code>-o beginning|end|stored|&lt;offset>|-&lt;offset></code>
-Offset to start consuming from.</p>
-
-<ul>
-<li><p><code>beginning</code> will start at the beginning.</p></li>
-<li><p><code>end</code> will start at the end.</p></li>
-<li><p><code>stored</code> will start at the last stored offset.</p></li>
-<li><p><code>&lt;offset></code> will start at absolute offset.</p></li>
-<li><p><code>-&lt;offset></code> will start at relative offset from the end.</p></li>
-</ul>
-</li>
-<li><p><code>-e</code>
-Exit consuming after last message received.</p></li>
-<li><p><code>-O</code>
-Print message offset using <code>-K</code> delimiter.</p></li>
-</ul>
-
-
-
-  <ol class='man-decor man-foot man foot'>
-    <li class='tl'></li>
-    <li class='tc'>January 2015</li>
-    <li class='tr'>kafkacat(1)</li>
-  </ol>
-
-  </div>
-</body>
-</html>
