+++ doc/kc.1
@@ -1,0 +1,181 @@
+.\" generated with Ronn/v0.7.3
+.\" http://github.com/rtomayko/ronn/tree/0.7.3
+.
+.TH "KC" "1" "January 2015" "" ""
+.
+.SH "NAME"
+\fBkc\fR \- generic producer and consumer for Apache Kafka
+.
+.SH "SYNOPSIS"
+.
+.nf
+
+kc producer [\-\-brokers=<brks>] [\-\-partition=<part>] [\-\-compression=<comp>]
+            [\-\-delimiter=<delim>] [\-\-key\-delimiter=<delim>] [\-\-count=<cnt>]
+            [\-T | \-\-tee] [\-q | \-\-quiet] [\-v | \-\-verbose] <topic> [<file>\|\.\|\.\|\.]
+
+kc consumer [\-\-brokers=<brks>] [\-\-partition=<part>] [\-e | \-\-exit]
+            [\-\-delimiter=<delim>] [\-\-key\-delimiter=<delim>] [\-\-count=<cnt>]
+            [\-q | \-\-quiet] [\-v | \-\-verbose] <topic>
+
+kc metadata [\-\-brokers=<brks>] [\-\-partition=<part>] [<topic>]
+
+kc \-\-help
+
+kc \-\-version
+.
+.fi
+.
+.SH "DESCRIPTION"
+\fBkc\fR also known as \fBkafkacat\fR is a generic non\-JVM producer and consumer for Apache Kafka 0\.8, think of it as a netcat for Kafka\.
+.
+.P
+In consumer mode, \fBkc\fR reads messages from a topic and partition and prints them to stdout using the configured message delimiter\.
+.
+.P
+In metadata mode, \fBkc\fR displays the current state of the Kafka cluster and its topics and partitions\.
+.
+.SH "COMMAND"
+.
+.IP "\[ci]" 4
+\fBproducer\fR reads messages from stdin, delimited with a configurable delimeter and produces them to the provided Kafka cluster, topic and partition\. See [PRODUCER] section for more information on producer\.
+.
+.IP "\[ci]" 4
+\fBconsumer\fR reads messages from a topic (and possibly a specific partition ) and prints them to stdout using the configured message delimiter\. See [CONSUMER] section for more information on consumer\.
+.
+.IP "\[ci]" 4
+\fBmetadata\fR displays the current state of the Kafka cluster and its topics and partitions\. See [METADATA] for more information on metadata\.
+.
+.IP "" 0
+.
+.SH "PRODUCER"
+.
+.SH "CONSUMER"
+.
+.SH "METADATA"
+.
+.SH "OPTIONS"
+.
+.SS "Generic options"
+.
+.IP "\[ci]" 4
+\fB\-b <brokers>, \-\-brokers=<brokers>\fR Comma separated list of broker(s) to bootstrap of the form "host[:port][,\|\.\|\.\|\.]", e\.g\. "host1:9292,host2:9293" [default: localhost]\.
+.
+.IP "\[ci]" 4
+\fB\-p <partition>, \-\-partition=<partition>\fR Send messages to a specific partition\. If \-1 is provided, a partition will be randomly uniformly selected for each message\.
+.
+.IP "\[ci]" 4
+\fB\-v, \-\-verbose\fR Augment verbosity level\.
+.
+.IP "\[ci]" 4
+\fB\-q, \-\-quiet\fR Quiet mode, do not report error messages to stderr\.
+.
+.IP "\[ci]" 4
+\fB\-V, \-\-version\fR Show the version on the first line and follow with the usage and help message\.
+.
+.IP "\[ci]" 4
+\fB\-h, \-\-help\fR Show the usage and help message\.
+.
+.IP "" 0
+.
+.SS "Producer options"
+.
+.IP "\[ci]" 4
+\fB\-d <delim>, \-\-delimiter=<delim>\fR Message delimiter character: a\-z\.\. | \er | \en | \et | \exNN [default: \en]\.
+.
+.IP "\[ci]" 4
+\fB\-k <delim>, \-\-key\-delimiter=<delim>\fR Key delimiter character: a\-z\.\. | \er | \en | \et | \exNN\.
+.
+.IP "\[ci]" 4
+\fB\-c <cnt>, \-\-count=<cnt>\fR Exit after producing \fB<cnt>\fR messages\.
+.
+.IP "\[ci]" 4
+\fB\-z <c>, \-\-compression=<c>\fR Compression applied to messages: snappy or gzip\.
+.
+.IP "\[ci]" 4
+\fB\-T, \-\-tee\fR Output queued messages to stdout, acting like tee\. Note that this does not guarantee that the message was correctly sent\.
+.
+.IP "" 0
+.
+.SS "Consumer options"
+.
+.IP "\[ci]" 4
+\fB\-d <delim>, \-\-delimiter=<delim>\fR Message delimiter character: a\-z\.\. | \er | \en | \et | \exNN [default: \en]\.
+.
+.IP "\[ci]" 4
+\fB\-k <delim>, \-\-key\-delimiter=<delim>\fR Key delimiter character: a\-z\.\. | \er | \en | \et | \exNN\.
+.
+.IP "\[ci]" 4
+\fB\-o <off>, \-\-offset=<off>\fR Exit after consumer \fB<cnt>\fR messages\.
+.
+.IP "\[ci]" 4
+\fB\-c <cnt>, \-\-count=<cnt>\fR Exit after consumer \fB<cnt>\fR messages\.
+.
+.IP "\[ci]" 4
+\fB\-e , \-\-exit\fR Exit when there is no message left to consumer in the topic\.
+.
+.IP "" 0
+.
+.SH "CONFIGURATION"
+.
+.SH "EXAMPLES"
+.
+.IP "\[ci]" 4
+Producing a single message
+.
+.IP "" 0
+.
+.IP "" 4
+.
+.nf
+
+    $ echo "test message" | kc producer test`
+.
+.fi
+.
+.IP "" 0
+.
+.IP "\[ci]" 4
+Consuming last message
+.
+.IP "" 0
+.
+.IP "" 4
+.
+.nf
+
+    $ kc consumer \-e \-o \-1 test
+    test message
+.
+.fi
+.
+.IP "" 0
+.
+.IP "\[ci]" 4
+Showing metadata on the test topic
+.
+.IP "" 0
+.
+.IP "" 4
+.
+.nf
+
+    $ kc metadata test
+    {
+      "brokers": [
+        { "id":0, "host":"localhost:9092" }
+      ],
+      "topics": [
+        {
+          "name": "test",
+          "partitions": [
+            { "id": 0, "leader": 0, "replicas": [0], "isrs": [0] }
+          ]
+        }
+      ]
+    }
+.
+.fi
+.
+.IP "" 0
+
